(#16188) quill: QUILL_X86ARCH should be optional
* quill: QUILL_X86ARCH should be optional

QUILL_X86ARCH uses some specific instructions for cache line optimisations (prefetch, clflushop) that are not available in each architecture. It also requires the user to compile using the executable by passing specific flags (march). This option won't be needed by the majority of the users. It also should be possible to disable. The default should be False and we should let the user explicitly enable it.

* quill: improve conditions when is_quilll_x86_arch is enabled